WIP refactoring.
